RD-981: Update demo app
