+++ Examples/MapTilerMobileDemo/app/src/main/java/com/maptilerdemo/maptilermobiledemo/HomeScreen.kt
@@ -8,9 +8,13 @@ package com.maptilerdemo.maptilermobiledemo
 
 import android.content.Context
 import android.util.Log
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import com.maptiler.maptilersdk.map.MTMapOptions
 import com.maptiler.maptilersdk.map.MTMapView
@@ -25,14 +29,25 @@ fun HomeScreen(
 ) {
     val mapController = MapController(context)
 
-    MTMapView(
-        MTMapReferenceStyle.STREETS,
-        MTMapOptions(),
-        mapController.controller,
-        modifier =
-            Modifier
-                .fillMaxSize(),
-    )
+    Box(modifier = Modifier.fillMaxSize()) {
+        MTMapView(
+            MTMapReferenceStyle.STREETS,
+            MTMapOptions(),
+            mapController.controller,
+            modifier =
+                Modifier
+                    .fillMaxSize(),
+        )
+
+        ZoomControl(
+            onZoomIn = { mapController.controller.zoomIn() },
+            onZoomOut = { mapController.controller.zoomOut() },
+            modifier =
+                Modifier
+                    .align(Alignment.TopEnd)
+                    .padding(20.dp),
+        )
+    }
 }
 
 class MapController(
